Fixed incorrect types in movie model

+++ LotrAPI.SDK/Models/Movie.cs
@@ -7,8 +7,8 @@ public record Movie(
     [property: JsonPropertyName("_id")] string Id,
     [property: JsonPropertyName("name")] string Name,
     [property: JsonPropertyName("runtimeInMinutes")] int RuntimeInMinutes,
-    [property: JsonPropertyName("budgetInMillions")] int BudgetInMillions,
-    [property: JsonPropertyName("boxOfficeRevenueInMillions")] int BoxOfficeRevenueInMillions,
+    [property: JsonPropertyName("budgetInMillions")] float BudgetInMillions,
+    [property: JsonPropertyName("boxOfficeRevenueInMillions")] float BoxOfficeRevenueInMillions,
     [property: JsonPropertyName("academyAwardNominations")] int AcademyAwardNominations,
-    [property: JsonPropertyName("rottenTomatoesScore")] int RottenTomatoesScore
+    [property: JsonPropertyName("rottenTomatoesScore")] float RottenTomatoesScore
 );
